Fixed bugs.
1. Reading a network file sometimes failed because of a parsing error.

2. Calling expf() in the neural net would sometimes overflow and not be
   handled properly, depending on the build flags.  We should just use
   exp(), and not std::exp(), instead.

Some consistency checks were added which should detect any remaining problems.

+++ mathfuncs.h
@@ -5,6 +5,6 @@
 
 inline float squash(float x)
 {
-    return 1 / (1 + expf(-x));
+    return  1 / (1 + exp(-x));
 }
